requirements.txt
Flask
requests
dnspython
fpdf2
